Harden SIM diagnostics and review workflows

+++ sim_control/focus_metrics.py
@@ -10,10 +10,10 @@ def sum_modified_laplacian(image: np.ndarray) -> float:
     array = np.asarray(image)
     if array.ndim != 2:
         raise ValueError("sum_modified_laplacian expects a 2-D image.")
-    data = array.astype(np.float64, copy=False)
+    data = array.astype(np.float32, copy=False)
     if min(data.shape) < 3:
         return 0.0
 
-    mlx = np.abs((2.0 * data[1:-1, 1:-1]) - data[1:-1, :-2] - data[1:-1, 2:])
-    mly = np.abs((2.0 * data[1:-1, 1:-1]) - data[:-2, 1:-1] - data[2:, 1:-1])
+    mlx = np.abs((2 * data[1:-1, 1:-1]) - data[1:-1, :-2] - data[1:-1, 2:])
+    mly = np.abs((2 * data[1:-1, 1:-1]) - data[:-2, 1:-1] - data[2:, 1:-1])
     return float(np.sum(mlx + mly))
